added PnP Core support to core library and its auth config to test console
