+++ M365AIMate.Console/M365AIMate.Console.csproj
@@ -7,11 +7,14 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AIPageMate.Core\M365AIMate.Core.csproj" />
+    <ProjectReference Include="..\M365AIMate.Core\M365AIMate.Core.csproj" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="7.0.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="7.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.1" />
+    <PackageReference Include="PnP.Core" Version="1.8.0" />
+    <PackageReference Include="PnP.Core.Auth" Version="1.8.0" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="appsettings.json">
