+++ M365AIMate.Console/appsettings.json
@@ -8,7 +8,51 @@
     }
   },
   "AllowedHosts": "*",
-  "clientId": "d59800b5-df28-4ab2-a192-860184c2133e",
-  "clientSecret": "zE38Q~J9TG4GWHF9EhKWKMSLnmi5rOg3O7TFyb5G",
-  "tenantId": "a1d5f937-b756-46d7-b92f-464629a6d985"
+  "clientId": "2e35359a-fc1f-4a0b-8fdb-310c3cd9415b",
+  "clientSecret": "0-a8Q~UgpnYTp7sur24DaZmxrZGob-HiCEoO9cot",
+  "tenantId": "d3ec8dc7-32e1-417b-a293-e77245357fb1",
+
+  "PnPCore": {
+    "DisableTelemetry": "false",
+    "HttpRequests": {
+      "UserAgent": "ISV|Contoso|ProductX",
+      "Timeout": "100",
+      "SharePointRest": {
+        "UseRetryAfterHeader": "false",
+        "MaxRetries": "10",
+        "DelayInSeconds": "3",
+        "UseIncrementalDelay": "true"
+      },
+      "MicrosoftGraph": {
+        "UseRetryAfterHeader": "true",
+        "MaxRetries": "10",
+        "DelayInSeconds": "3",
+        "UseIncrementalDelay": "true"
+      }
+    },
+    "PnPContext": {
+      "GraphFirst": "true",
+      "GraphCanUseBeta": "true",
+      "GraphAlwaysUseBeta": "false"
+    },
+    "Sites": {
+      "SiteToWorkWith": {
+        "SiteUrl": "https://tgodev.sharepoint.com/"
+      }
+    },
+    "Credentials": {
+      "DefaultConfiguration": "x509Certificate",
+      "Configurations": {
+        "x509Certificate": {
+          "ClientId": "7055fd02-5832-4215-81aa-ce3449197c4e",
+          "TenantId": "d3ec8dc7-32e1-417b-a293-e77245357fb1",
+          "X509Certificate": {
+            "StoreName": "My",
+            "StoreLocation": "CurrentUser",
+            "ThumbPrint": "5C46D1A0AEAC7DFD23586F3ECE4F34EE4102DD18"
+          }
+        }
+      }
+    }
+  }
 }
