+++ M365AIMate.Core/M365AIMate.Core.csproj
@@ -10,5 +10,7 @@
     <PackageReference Include="Microsoft.Graph.Core" Version="3.0.0" />
     <PackageReference Include="Microsoft.Graph.Beta" Version="5.20.0-preview" />
     <PackageReference Include="Azure.Identity" Version="1.8.2" />
+    <PackageReference Include="PnP.Core" Version="1.8.0" />
+    <PackageReference Include="PnP.Core.Admin" Version="1.8.0" />
   </ItemGroup>
 </Project>
