+++ M365AIMate.TeamsTab/M365AIMate.TeamsTab.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net7.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
@@ -20,6 +20,10 @@
     <PackageReference Include="Microsoft.Graph" Version="4.44.0" />
     <PackageReference Include="Microsoft.Fast.Components.FluentUI" Version="1.5.3" />
     <PackageReference Include="Microsoft.TeamsFx" Version="1.2.*" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\M365AIMate.Core\M365AIMate.Core.csproj" />
   </ItemGroup>
 
 </Project>
